% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{backup_archives}
\alias{backup_archives}
\title{Backup the current archive files}
\usage{
backup_archives(archives = DEFAULT_ARCHIVES)
}
\arguments{
\item{archives}{The name of the folder to use for the archives.}
}
\description{
Backup the current archive files
}
