% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clbrief.R
\docType{package}
\name{clbrief}
\alias{clbrief}
\alias{clbrief-package}
\title{clbrief: A package for downloading data from the Research Briefings API}
\description{
The Research Briefings API contains data on all research briefings published
at researchbriefings.parliament.uk. This package provides functions for downloading data
on briefings.
}
