% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{create_archives}
\alias{create_archives}
\title{Create initial archive files}
\usage{
create_archives(archives = DEFAULT_ARCHIVES)
}
\arguments{
\item{archives}{The name of the folder to use for the archives.}
}
\description{
Create initial archive files
}
