% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_db}
\alias{create_db}
\title{Create database from csvs}
\usage{
create_db(dbfile = BRIEFINGS_DB, dbdir = DATABASE_DIR)
}
\arguments{
\item{dbfile}{Path to create a SQLite briefings database.}

\item{dbdir}{Path to the directory containing the csvs to import.}
}
\description{
Create database from csvs
}
