% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.R
\name{create_mirror}
\alias{create_mirror}
\title{Create a mirror from csvs}
\usage{
create_mirror(dbfile = MIRROR_DB, dbdir = DATABASE_DIR)
}
\arguments{
\item{dbfile}{Path to create a SQLite briefings database.}

\item{dbdir}{Path to the directory containing the csvs to import.}
}
\description{
Create a mirror from csvs
}
