% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{extract_db}
\alias{extract_db}
\title{Extract database to csvs}
\usage{
extract_db(dbfile = BRIEFINGS_DB, dbdir = DATABASE_DIR)
}
\arguments{
\item{dbfile}{Path to a SQLite briefings database.}

\item{dbdir}{Path to the directory containing the csvs to import.}
}
\description{
Extract database to csvs
}
