% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.R
\name{extract_mirror}
\alias{extract_mirror}
\title{Extract a mirror to csvs}
\usage{
extract_mirror(dbfile = MIRROR_DB, dbdir = DATABASE_DIR)
}
\arguments{
\item{dbfile}{Path to a SQLite briefings database.}

\item{dbdir}{Path to the directory containing the csvs to import.}
}
\description{
Extract a mirror to csvs
}
