% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_all_json}
\alias{fetch_all_json}
\title{Fetch all json data for a set of briefings}
\usage{
fetch_all_json(resources, pause = 1)
}
\arguments{
\item{resources}{A vector of resource urls for a set of briefing papers.
These are the urls stored in the first column of the dataframe returned by
\code{get_briefings} and \code{fetch_briefings}.}

\item{pause}{The number of seconds to pause between page requests.}
}
\value{
A list of briefings as json.
}
\description{
\code{fetch_all_json} gets all of the available data for a set of briefings.
This provides some data that is not available in the main research briefings
feed and must be downloaded separately from the API for each individual
briefing paper. This function will therefore make multiple calls to the API.
}
