% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_briefings}
\alias{fetch_briefings}
\title{Fetch summary data on briefings and return as a tibble}
\usage{
fetch_briefings(pages = 1, pagesize = 500, pause = 1)
}
\arguments{
\item{pages}{The number of pages to pull from the API. The default is 1.}

\item{pagesize}{The number of briefings per page. The default is 500 and the
maximum allowed value is 500.}

\item{pause}{The number of seconds to pause between page requests.}
}
\value{
A tibble of sumamry data on briefings.
}
\description{
\code{fetch_briefings} fetches data on all briefings up to a given
number derived from the number of pages of results and the pagesize and
returns summary data on each briefing.
}
\details{
Note that data which has a one to many realtionship with a briefing is not
included in the tibble returned by this function. See the the other
functions for extracting nested data.
}
