% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_briefings_json}
\alias{fetch_briefings_json}
\title{Fetch data on all briefings up to a given number as JSON}
\usage{
fetch_briefings_json(pages = 1, pagesize = 500, pause = 1)
}
\arguments{
\item{pages}{The number of pages to pull from the API. The default is 1.}

\item{pagesize}{The number of briefings per page. The default is 500 and the
maximum allowed value is 500.}

\item{pause}{The number of seconds to pause between page requests.}
}
\value{
A list of briefings as JSON.
}
\description{
\code{fetch_briefings_json} fetches data on all briefings up to a given
number derived from the number of pages of results and the pagesize.
}
