% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_sections}
\alias{fetch_sections}
\title{Fetch data on the sections that produced a briefing and return as a tibble}
\usage{
fetch_sections(pages = 1, pagesize = 500, pause = 1)
}
\arguments{
\item{pages}{The number of pages to pull from the API. The default is 1.}

\item{pagesize}{The number of briefings per page. The default is 500 and the
maximum allowed value is 500.}

\item{pause}{The number of seconds to pause between page requests.}
}
\value{
A tibble of data on briefings and their sections in long form.
}
\description{
\code{fetch_sections} fetches data on all briefings up to a given number
derived from the number of pages of results and the pagesize and returns
the sections responsible for producing each briefing.
}
