% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_db_authors}
\alias{get_db_authors}
\title{Get the authors table from the database as a tibble}
\usage{
get_db_authors(dbc = NULL, dbfile = BRIEFINGS_DB)
}
\arguments{
\item{dbc}{A connection to a SQLite briefings database.}

\item{dbfile}{Path to a SQLite briefings database.}
}
\description{
Get the authors table from the database as a tibble
}
