% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_db_table}
\alias{get_db_table}
\title{Get a table from the database as a tibble}
\usage{
get_db_table(table_name, dbc = NULL, dbfile = BRIEFINGS_DB)
}
\arguments{
\item{table_name}{The name of the table to retreive from the database.}

\item{dbc}{A connection to a SQLite briefings database.}

\item{dbfile}{Path to a SQLite briefings database.}
}
\description{
Get a table from the database as a tibble
}
