% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_db_topics}
\alias{get_db_topics}
\title{Get the topics table from the database as a tibble}
\usage{
get_db_topics(dbc = NULL, dbfile = BRIEFINGS_DB)
}
\arguments{
\item{dbc}{A connection to a SQLite briefings database.}

\item{dbfile}{Path to a SQLite briefings database.}
}
\description{
Get the topics table from the database as a tibble
}
