% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{get_documents}
\alias{get_documents}
\title{Get a tibble of documents for each briefing}
\usage{
get_documents(briefings_all_json)
}
\arguments{
\item{briefings_all_json}{A list of briefings as returned by
\code{fetch_all_json}}
}
\value{
A tibble of data on briefings and their documents in long form.
}
\description{
\code{get_documents} takes a list of briefings returned from the API as JSON
and returns a tibble of documents for those briefings, with one row per
unqiue combination of briefing and document. Data on documents is not
available from the main research breifings feed so this function requires
the json produced from \code{fetch_all_json}.
}
