% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.R
\name{get_mirror_briefings}
\alias{get_mirror_briefings}
\title{Get the briefings table from the mirror as a tibble}
\usage{
get_mirror_briefings(dbc = NULL, dbfile = MIRROR_DB)
}
\arguments{
\item{dbc}{A connection to a SQLite briefings database.}

\item{dbfile}{Path to a SQLite briefings database.}
}
\description{
Get the briefings table from the mirror as a tibble
}
