% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{get_sections}
\alias{get_sections}
\title{Get a tibble of sections for each briefing}
\usage{
get_sections(briefings_json)
}
\arguments{
\item{briefings}{A list of briefings as returned by
\code{fetch_briefings_json}}
}
\value{
A tibble of data on briefings and their section in long form.
}
\description{
\code{get_sections} takes a list of briefings returned from the API as JSON
and returns a tibble of authors for those briefings, with one row per unqiue
combination of briefing and author.
}
