% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{get_topics}
\alias{get_topics}
\title{Get a tibble of topics for each briefing}
\usage{
get_topics(briefings_json)
}
\arguments{
\item{briefings_json}{A list of briefings as returned by
\code{fetch_briefings_json}}
}
\value{
A tibble of data on briefings and their topics in long form.
}
\description{
\code{get_topics} takes a list of briefings returned from the API
as JSON and returns a tibble of topics for those briefings, with one row per
unqiue combination of briefing and topic.
}
