% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{update_archives}
\alias{update_archives}
\title{Update the current archive files}
\usage{
update_archives(archives = DEFAULT_ARCHIVES)
}
\arguments{
\item{archives}{The name of the folder to use for the archives.}
}
\description{
Update the current archive files
}
