% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{update_db}
\alias{update_db}
\title{Update the database}
\usage{
update_db(dbfile = BRIEFINGS_DB, backup_dbfile = BACKUP_DB)
}
\arguments{
\item{dbfile}{Path to a SQLite briefings database.}

\item{backup_dbfile}{Path to a backup SQLite briefings database.}
}
\description{
Update the database
}
