% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{update_db_briefings}
\alias{update_db_briefings}
\title{Update the briefings table with new data from the api}
\usage{
update_db_briefings(briefings_data, dbc = NULL, dbfile = BRIEFINGS_DB)
}
\arguments{
\item{briefings_data}{Data on briefings returned from \code{get_briefings}.}

\item{dbc}{A connection to a SQLite briefings database.}

\item{dbfile}{Path to a SQLite briefings database.}
}
\description{
Update the briefings table with new data from the api
}
