% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{update_db_documents}
\alias{update_db_documents}
\title{Update the documents table with new data from the api}
\usage{
update_db_documents(documents_data, dbc = NULL, dbfile = BRIEFINGS_DB)
}
\arguments{
\item{documents_data}{Data on documents returned from \code{get_documents}.}

\item{dbc}{A connection to a SQLite briefings database.}

\item{dbfile}{Path to a SQLite briefings database.}
}
\description{
Update the documents table with new data from the api
}
