% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{update_db_topics}
\alias{update_db_topics}
\title{Update the topics table with new data from the api}
\usage{
update_db_topics(topics_data, dbc = NULL, dbfile = BRIEFINGS_DB)
}
\arguments{
\item{topics_data}{Data on topics returned from \code{get_topics}.}

\item{dbc}{A connection to a SQLite briefings database.}

\item{dbfile}{Path to a SQLite briefings database.}
}
\description{
Update the topics table with new data from the api
}
