% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.R
\name{update_mirror}
\alias{update_mirror}
\title{Update the mirror}
\usage{
update_mirror(dbfile = MIRROR_DB, backup_dbfile = MIRROR_BACKUP_DB)
}
\arguments{
\item{dbfile}{Path to a SQLite briefings database.}

\item{backup_dbfile}{Path to a backup SQLite briefings database.}
}
\description{
Update the mirror
}
