% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.R
\name{update_mirror_authors}
\alias{update_mirror_authors}
\title{Update the authors table with new data from the api}
\usage{
update_mirror_authors(authors_data, dbc = NULL, dbfile = MIRROR_DB)
}
\arguments{
\item{authors_data}{Data on authors returned from \code{get_authors}.}

\item{dbc}{A connection to a SQLite briefings database.}

\item{dbfile}{Path to a SQLite briefings database.}
}
\description{
Update the authors table with new data from the api
}
