% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror.R
\name{update_mirror_sections}
\alias{update_mirror_sections}
\title{Update the sections table with new data from the api}
\usage{
update_mirror_sections(sections_data, dbc = NULL, dbfile = MIRROR_DB)
}
\arguments{
\item{sections_data}{Data on sections returned from \code{get_sections}.}

\item{dbc}{A connection to a SQLite briefings database.}

\item{dbfile}{Path to a SQLite briefings database.}
}
\description{
Update the sections table with new data from the api
}
